Move formatter tests in a dedicated file
